auth/code/requirement.txt
requests
jsonify
boto3
python-jose
flask
gunicorn